fs: move SyncWriteStream to internal/fs
Move the implementation of SyncWriteStream to internal/fs.
