+++ lib/fs.js
@@ -14,6 +14,9 @@ const Stream = require('stream').Stream;
 const EventEmitter = require('events');
 const FSReqWrap = binding.FSReqWrap;
 const FSEvent = process.binding('fs_event_wrap').FSEvent;
+const internalFS = require('internal/fs');
+const assertEncoding = internalFS.assertEncoding;
+const SyncWriteStream = internalFS.SyncWriteStream;
 
 Object.defineProperty(exports, 'constants', {
   configurable: false,
@@ -94,12 +97,6 @@ function makeCallback(cb) {
   return function() {
     return cb.apply(null, arguments);
   };
-}
-
-function assertEncoding(encoding) {
-  if (encoding && !Buffer.isEncoding(encoding)) {
-    throw new Error('Unknown encoding: ' + encoding);
-  }
 }
 
 function nullCheck(path, callback) {
@@ -2144,75 +2141,12 @@ WriteStream.prototype.close = ReadStream.prototype.close;
 // There is no shutdown() for files.
 WriteStream.prototype.destroySoon = WriteStream.prototype.end;
 
-
 // SyncWriteStream is internal. DO NOT USE.
-// Temporary hack for process.stdout and process.stderr when piped to files.
-function SyncWriteStream(fd, options) {
-  Stream.call(this);
-
-  options = options || {};
-
-  this.fd = fd;
-  this.writable = true;
-  this.readable = false;
-  this.autoClose = options.autoClose === undefined ? true : options.autoClose;
-}
-
-util.inherits(SyncWriteStream, Stream);
-
-
-// Export
+// todo(jasnell): "Docs-only" deprecation for now. This was never documented
+// so there's no documentation to modify. In the future, add a runtime
+// deprecation.
 Object.defineProperty(fs, 'SyncWriteStream', {
   configurable: true,
   writable: true,
   value: SyncWriteStream
 });
-
-SyncWriteStream.prototype.write = function(data, arg1, arg2) {
-  var encoding, cb;
-
-  // parse arguments
-  if (arg1) {
-    if (typeof arg1 === 'string') {
-      encoding = arg1;
-      cb = arg2;
-    } else if (typeof arg1 === 'function') {
-      cb = arg1;
-    } else {
-      throw new Error('Bad arguments');
-    }
-  }
-  assertEncoding(encoding);
-
-  // Change strings to buffers. SLOW
-  if (typeof data === 'string') {
-    data = Buffer.from(data, encoding);
-  }
-
-  fs.writeSync(this.fd, data, 0, data.length);
-
-  if (cb) {
-    process.nextTick(cb);
-  }
-
-  return true;
-};
-
-
-SyncWriteStream.prototype.end = function(data, arg1, arg2) {
-  if (data) {
-    this.write(data, arg1, arg2);
-  }
-  this.destroy();
-};
-
-
-SyncWriteStream.prototype.destroy = function() {
-  if (this.autoClose)
-    fs.closeSync(this.fd);
-  this.fd = null;
-  this.emit('close');
-  return true;
-};
-
-SyncWriteStream.prototype.destroySoon = SyncWriteStream.prototype.destroy;
