+++ lib/internal/process/stdio.js
@@ -145,7 +145,7 @@ function createWritableStdioStream(fd) {
       break;
 
     case 'FILE':
-      const fs = require('fs');
+      const fs = require('internal/fs');
       stream = new fs.SyncWriteStream(fd, { autoClose: false });
       stream._type = 'fs';
       break;
